all: minor refactoring
An older refactoring moved toSizedDeltas and fromSizedDeltas from section.go to bits.go
but forgot to move the test. Cleaning this up here.

Also fixed a couple of typos, deleted stale comments etc.
