+++ bits_test.go
@@ -108,6 +108,15 @@ func TestNextFileIndex(t *testing.T) {
 	}
 }
 
+func TestDeltas(t *testing.T) {
+	in := []uint32{1, 72, 0xfff}
+	out := toSizedDeltas(in)
+	round := fromSizedDeltas(out, nil)
+	if !reflect.DeepEqual(in, round) {
+		t.Errorf("got %v, want %v", round, in)
+	}
+}
+
 func TestSizedDeltas(t *testing.T) {
 	encode := func(nums []uint32) []byte {
 		return toSizedDeltas(nums)
